Fix failing 3.8 test and lint

+++ tests/unit/cag/test_one_hot_encoding.py
@@ -137,7 +137,7 @@ class TestOneHotEncoding:
             'table': {
                 'a': np.dtype('float'),
                 'b': np.dtype('float'),
-                'c': np.dtype('int')
+                'c': np.dtype('int64'),
             }
         }
